[FEAT] PointCut Before 메서드 이용하게 변경 (#99)
